backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pymongo
monogoengine
django-rest-framework-mongoengine
dotenv